+++ app/src/main/java/edu/stanford/cardinalkit/di/AppModule.kt
@@ -26,9 +26,10 @@ import edu.stanford.cardinalkit.data.repositories.SurveyRepositoryImpl
 import edu.stanford.cardinalkit.domain.repositories.AuthRepository
 import edu.stanford.cardinalkit.domain.repositories.ContactsRepository
 import edu.stanford.cardinalkit.domain.repositories.SurveyRepository
-import edu.stanford.cardinalkit.domain.use_cases.GetContacts
-import edu.stanford.cardinalkit.domain.use_cases.UploadSurvey
-import edu.stanford.cardinalkit.domain.use_cases.UseCases
+import edu.stanford.cardinalkit.domain.use_cases.contacts.ContactsUseCases
+import edu.stanford.cardinalkit.domain.use_cases.contacts.GetContacts
+import edu.stanford.cardinalkit.domain.use_cases.surveys.SurveysUseCases
+import edu.stanford.cardinalkit.domain.use_cases.surveys.UploadSurvey
 import javax.inject.Named
 
 @Module
@@ -135,14 +136,20 @@ class AppModule {
     fun provideContactsRepository(context: Context): ContactsRepository = ContactsRepositoryImpl(context)
 
     @Provides
-    @Named(Constants.USE_CASES)
-    fun provideUseCases(
+    @Named(Constants.SURVEYS_USE_CASES)
+    fun provideSurveysUseCases(
         @Named(Constants.SURVEY_REPOSITORY)
         surveyRepository: SurveyRepository,
+    ) = SurveysUseCases(
+        uploadSurvey = UploadSurvey(surveyRepository)
+    )
+
+    @Provides
+    @Named(Constants.CONTACTS_USE_CASES)
+    fun provideContactsUseCases(
         @Named(Constants.CONTACTS_REPOSITORY)
         contactsRepository: ContactsRepository
-    ) = UseCases(
-        uploadSurvey = UploadSurvey(surveyRepository),
+    ) = ContactsUseCases(
         getContacts = GetContacts(contactsRepository)
     )
 }
