--- app/src/main/java/edu/stanford/cardinalkit/domain/use_cases/UseCases.kt
@@ -1,6 +1,0 @@
-package edu.stanford.cardinalkit.domain.use_cases
-
-data class UseCases(
-    val uploadSurvey: UploadSurvey,
-    val getContacts: GetContacts
-)
