+++ app/src/main/java/edu/stanford/cardinalkit/domain/use_cases/contacts/ContactsUseCases.kt
@@ -1,0 +1,5 @@
+package edu.stanford.cardinalkit.domain.use_cases.contacts
+
+data class ContactsUseCases (
+    val getContacts: GetContacts
+)
