+++ app/src/main/java/edu/stanford/cardinalkit/domain/use_cases/contacts/GetContacts.kt
@@ -1,4 +1,4 @@
-package edu.stanford.cardinalkit.domain.use_cases
+package edu.stanford.cardinalkit.domain.use_cases.contacts
 
 import edu.stanford.cardinalkit.domain.repositories.ContactsRepository
