+++ app/src/main/java/edu/stanford/cardinalkit/domain/use_cases/surveys/SurveysUseCases.kt
@@ -1,0 +1,7 @@
+package edu.stanford.cardinalkit.domain.use_cases.surveys
+
+import edu.stanford.cardinalkit.domain.use_cases.surveys.UploadSurvey
+
+data class SurveysUseCases(
+    val uploadSurvey: UploadSurvey
+)
