+++ app/src/main/java/edu/stanford/cardinalkit/domain/use_cases/surveys/UploadSurvey.kt
@@ -1,4 +1,4 @@
-package edu.stanford.cardinalkit.domain.use_cases
+package edu.stanford.cardinalkit.domain.use_cases.surveys
 
 import edu.stanford.cardinalkit.domain.repositories.SurveyRepository
