+++ app/src/main/java/edu/stanford/cardinalkit/presentation/surveys/SurveyViewModel.kt
@@ -7,14 +7,14 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import edu.stanford.cardinalkit.common.Constants
 import edu.stanford.cardinalkit.domain.models.Response
-import edu.stanford.cardinalkit.domain.use_cases.UseCases
+import edu.stanford.cardinalkit.domain.use_cases.surveys.SurveysUseCases
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 import javax.inject.Named
 
 @HiltViewModel
 class SurveyViewModel @Inject constructor(
-    @Named(Constants.USE_CASES) private val useCases: UseCases
+    @Named(Constants.SURVEYS_USE_CASES) private val useCases: SurveysUseCases
 ): ViewModel() {
 
     private val _surveyUploadedState = mutableStateOf<Response<Void?>>(Response.Success(null))
